+++ apps/docs/src/lib/config.ts
@@ -7,5 +7,5 @@ export type DocumentSiteConfiguration = {
 export const config: DocumentSiteConfiguration = {
   showDrafts: process.env.SHOW_DRAFTS === 'true',
   feedbackFormUrl: process.env.FEEDBACK_FORM_URL || '',
-  signUpFormUrl: process.env.SIGN_UP_FORM_URL || '',
+  signUpFormUrl: process.env.SIGNUP_FORM_URL || '',
 };