+++ packages/react/ds/src/progress-bar/progress-bar.test.tsx
@@ -10,7 +10,6 @@ describe('ProgressBar', () => {
   it('should render a progress bar with a specified value', () => {
     const screen = renderProgressBar({
       value: 50,
-      size: 'md',
     });
 
     const progressBar = screen.getByTestId('progress-bar');
@@ -19,13 +18,11 @@ describe('ProgressBar', () => {
     const progressBarInner = progressBar.firstElementChild;
     expect(progressBarInner).toBeTruthy();
     expect(progressBarInner).toHaveStyle('width: 50%');
-    expect(progressBarInner).toHaveClass('gi-bg-gray-700', 'gi-h-4');
   });
 
   it('should render an indeterminate progress bar correctly', () => {
     const screen = renderProgressBar({
       isIndeterminate: true,
-      size: 'lg',
     });
 
     const progressBar = screen.getByTestId('progress-bar');
@@ -33,18 +30,13 @@ describe('ProgressBar', () => {
 
     const progressBarInner = progressBar.firstElementChild;
     expect(progressBarInner).toBeTruthy();
-    expect(progressBarInner).toHaveClass(
-      'gi-progress-bar-indeterminate',
-      'gi-bg-gray-700',
-      'gi-h-6',
-    );
+    expect(progressBarInner).toHaveClass('gi-progress-bar-indeterminate');
   });
 
   it('should render progress bar with final value set', () => {
     const screen = renderProgressBar({
       max: 500,
       value: 200,
-      size: 'lg',
     });
 
     const progressBar = screen.getByTestId('progress-bar');
@@ -58,7 +50,6 @@ describe('ProgressBar', () => {
   it('should pass axe accessibility tests', async () => {
     const screen = renderProgressBar({
       value: 75,
-      size: 'sm',
     });
 
     await screen.axe();