+++ packages/react/ds/src/progress-stepper/progress-stepper.tsx
@@ -1,77 +1,98 @@
 import { Icon } from '../icon/icon.js';
-import type {
-  ConnectorProps,
-  ProgressStepperProps,
-  StepProps,
+import {
+  ProgressStepperIndicator,
+  type ConnectorProps,
+  type ProgressStepperIndicatorType,
+  type ProgressStepperProps,
+  type StepProps,
 } from './types.js';
 
 const Connector = ({
-  stepNumber,
   isNextStep,
   orientation = 'horizontal',
+  isCurrentStep,
+  isCompleted,
 }: ConnectorProps) => {
-  if (stepNumber > 1) {
-    return (
-      <div
-        data-orientation={orientation}
-        data-next={isNextStep}
-        className="gi-progress-stepper-step-connector"
-        aria-hidden="true"
-      >
-        <span />
-      </div>
-    );
-  }
+  return (
+    <div
+      data-orientation={orientation}
+      data-next={isNextStep}
+      data-completed={isCompleted}
+      data-current={isCurrentStep}
+      className="gi-progress-stepper-step-connector"
+      aria-hidden="true"
+    >
+      <span />
+      {isCurrentStep ? <span /> : null}
+    </div>
+  );
+};
 
-  return null;
+const getIndicatorClasses = (indicator: ProgressStepperIndicatorType) => {
+  const indicatorClasses = {
+    [ProgressStepperIndicator.hashtag]: {
+      completed: <Icon icon="check" />,
+      current: '#',
+      next: '#',
+    },
+  };
+
+  return indicatorClasses[indicator];
 };
 
 const Step = ({
-  children,
   isCurrentStep,
   isCompleted,
+  isLastStep,
   stepNumber,
   orientation,
+  children,
+  indicator = 'hashtag',
 }: StepProps) => {
   const isNextStep = !isCompleted && !isCurrentStep;
-  const isVertical = orientation === 'vertical';
-  const isHorizontal = orientation === 'horizontal';
+  const { current, completed, next } = getIndicatorClasses(
+    indicator || ProgressStepperIndicator.hashtag,
+  );
+
+  const getProgressIconStep = () => {
+    if (isCompleted) {
+      return completed;
+    } else if (isCurrentStep) {
+      return current;
+    }
+    return next;
+  };
 
   return (
-    <div className="gi-flex gi-relative gi-flex-1 gi-min-h-20 gi-min-w-20">
+    <div className="gi-relative">
       <div
         className="gi-progress-stepper-step-container"
         data-orientation={orientation}
         data-current={isCurrentStep}
         data-completed={isCompleted}
         data-next={isNextStep}
+        data-indicator={indicator}
         role="listitem"
         aria-labelledby={`step-label-${stepNumber}`}
       >
-        <div className="gi-progress-stepper-step gi-relative">
-          {isCompleted ? <Icon icon="check" /> : <div>{stepNumber}</div>}
-          {isVertical ? (
-            <Connector
-              isNextStep={isNextStep}
-              orientation={orientation}
-              stepNumber={stepNumber}
-            />
-          ) : null}
-        </div>
+        <div className="gi-progress-stepper-step">{getProgressIconStep()}</div>
         <div
           className="gi-progress-stepper-step-label"
+          data-orientation={orientation}
           id={`step-label-${stepNumber}`}
         >
           {children}
         </div>
       </div>
-      {isHorizontal ? (
+      {isLastStep ? null : (
         <Connector
+          isCurrentStep={isCurrentStep}
           isNextStep={isNextStep}
+          isCompleted={isCompleted}
           orientation={orientation}
           stepNumber={stepNumber}
         />
-      ) : null}
+      )}
     </div>
   );
 };
@@ -90,20 +111,28 @@ export const ProgressStepper = ({
       role="list"
       aria-live="polite"
     >
-      {steps.map((step, index) => (
-        <Step
-          key={`progress-stepper-step-${index}`}
-          stepNumber={index + 1}
-          isCurrentStep={!completeAll && currentStepIndex === index}
-          isCompleted={
-            completeAll ||
-            (index < currentStepIndex && index !== currentStepIndex)
-          }
-          orientation={orientation}
-        >
-          {step}
-        </Step>
-      ))}
+      {steps.map((step, index) => {
+        const [isCurrentStep, isLastStep, isCompleted] = [
+          !completeAll && currentStepIndex === index,
+          index === steps.length - 1,
+          completeAll ||
+            (index < currentStepIndex && index !== currentStepIndex),
+        ];
+        return (
+          <div className="gi-w-full">
+            <Step
+              key={`progress-stepper-step-${index}`}
+              stepNumber={index + 1}
+              isCurrentStep={isCurrentStep}
+              isCompleted={isCompleted}
+              orientation={orientation}
+              isLastStep={isLastStep}
+            >
+              {step}
+            </Step>
+          </div>
+        );
+      })}
     </div>
   );
 };