+++ packages/react/ds/src/progress-stepper/types.ts
@@ -1,5 +1,12 @@
 type Orientation = 'vertical' | 'horizontal';
 
+export const ProgressStepperIndicator = {
+  hashtag: 'hashtag',
+} as const;
+
+export type ProgressStepperIndicatorType =
+  (typeof ProgressStepperIndicator)[keyof typeof ProgressStepperIndicator];
+
 export type ProgressStepperProps = {
   steps: string[];
   currentStepIndex?: number;
@@ -10,13 +17,17 @@ export type ProgressStepperProps = {
 export type StepProps = {
   children: string;
   isCurrentStep: boolean;
+  isLastStep: boolean;
   isCompleted: boolean;
   stepNumber: number;
   orientation?: Orientation;
+  indicator?: ProgressStepperIndicatorType;
 };
 
 export type ConnectorProps = {
   stepNumber: number;
   isNextStep: boolean;
+  isCurrentStep: boolean;
+  isCompleted: boolean;
   orientation?: Orientation;
 };